requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pytest>=7.4.0
httpx>=0.25.0
pydantic>=2.6.0
python-multipart>=0.0.6
openpyxl>=3.1.2
qrcode[pil]>=7.4.2
reportlab>=4.1.0
google-genai>=0.1.0
python-docx>=1.0.0
